+++ src/core/marker.ts
@@ -3,30 +3,30 @@ import { createComment } from './utilities/create_comment';
 
 /** Base class for markers. */
 export interface Marker {
-  readonly id: number;
+  readonly _id: string;
   readonly value: unknown;
 
   toString(): string;
 }
 
 export class AttributeMarker implements Marker {
-  readonly id: number = uid();
+  readonly _id: string = uid();
 
   constructor(readonly value: unknown) {}
 
   toString(): string {
-    return `{{${this.id}}}`;
+    return `{{${this._id}}}`;
   }
 }
 
 export class NodeMarker implements Marker {
-  readonly id: number = uid();
+  readonly _id: string = uid();
 
   constructor(readonly value: unknown) {}
 
   toString(): string {
     // Here we just define place for nodes.
-    return createComment(`{{${this.id}}}`);
+    return createComment(`{{${this._id}}}`);
   }
 }