+++ src/core/reactive.ts
@@ -1,3 +1,4 @@
+import { uid } from './utilities/uid';
 import { bridges } from './bridge';
 import { Action, Dependency } from './dependency';
 
@@ -18,6 +19,10 @@ type Reactive<T extends object> = {
  * change - provide a value.
  */
 export function reactive<T extends object>(obj: T): Reactive<T> {
+  // Store reactive's id to touch only those bridges,
+  // that have dependencies from _obj_'s properties.
+  const reactiveId = uid();
+
   return new Proxy<Reactive<T>>((obj as unknown) as Reactive<T>, {
     get(target: Reactive<T>, property: string, receiver: unknown) {
       return <R>(value: Action<T[keyof T], R> | T[keyof T]) => {
@@ -30,6 +35,7 @@ export function reactive<T extends object>(obj: T): Reactive<T> {
 
           if (typeof value === 'function') {
             return new Dependency<T[keyof T], R>(
+              reactiveId,
               property,
               value as Action<T[keyof T], R>,
               valueInState
@@ -38,7 +44,11 @@ export function reactive<T extends object>(obj: T): Reactive<T> {
             if (!Object.is(value, valueInState)) {
               Reflect.set(target, property, value, receiver);
               bridges
-                .filter((bridge) => bridge.dependency.property === property)
+                .filter(
+                  (bridge) =>
+                    bridge.dependency._id === reactiveId &&
+                    bridge.dependency._property === property
+                )
                 .forEach((bridge) => bridge.update(value));
             }
           }