+++ src/core/utilities/uid.ts
@@ -1,4 +1,4 @@
 /** Generates unique id. */
-export function uid(): number {
-  return window.crypto.getRandomValues(new Uint32Array(1))[0];
+export function uid(): string {
+  return window.crypto.getRandomValues(new Uint32Array(1))[0].toString(16);
 }